+++ lib/inject.js
@@ -6,7 +6,7 @@ let lightMyRequest
 // HTTP injection handling
 // If the server is not ready yet, this
 // utility will automatically force it.
-module.exports = function createInject (httpHandler) {
+module.exports = function createInjectFunction (httpHandler) {
   return function inject (opts, cb) {
     // lightMyRequest is dynamically loaded as it seems very expensive
     // because of Ajv
